[pink_balancer] Rename WBC to JointController
