+++ agents/pink_balancer/pink_balancer.py
@@ -31,7 +31,7 @@ import yaml
 from loop_rate_limiters import AsyncRateLimiter
 from vulp.spine import SpineInterface
 
-from agents.pink_balancer.whole_body_controller import WholeBodyController
+from agents.pink_balancer.joint_controller import JointController
 from utils.realtime import configure_cpu
 from utils.spdlog import logging
 
@@ -77,9 +77,7 @@ async def run(
         config: Configuration dictionary.
         frequency: Control frequency in Hz.
     """
-    whole_body_controller = WholeBodyController(
-        config, visualize=args.visualize
-    )
+    joint_controller = JointController(config, visualize=args.visualize)
     debug: Dict[str, Any] = {}
     dt = 1.0 / frequency
     rate = AsyncRateLimiter(frequency, "controller")
@@ -87,7 +85,7 @@ async def run(
     observation = spine.get_observation()  # pre-reset observation
     while True:
         observation = spine.get_observation()
-        action = whole_body_controller.cycle(observation, dt)
+        action = joint_controller.cycle(observation, dt)
         action_time = time.time()
         spine.set_action(action)
         debug["rate"] = {
